refactor: restructure tests to match src
