+++ tests/library_analyzer/processing/annotations/test_generate_annotations.py
