+++ tests/library_analyzer/processing/api/test_instance_attributes.py
