+++ tests/library_analyzer/processing/dependencies/test_get_dependency.py
