+++ tests/migration/annotations/test_enum_migration.py
