+++ tests/migration/annotations/test_remove_migration.py
