+++ tests/migration/test_migration.py
@@ -16,14 +16,14 @@ from library_analyzer.processing.migration.model import (
     Mapping,
     SimpleDiffer,
 )
-from tests.processing.migration.annotations.test_boundary_migration import (
+from tests.migration.annotations.test_boundary_migration import (
     migrate_boundary_annotation_data_duplicated,
     migrate_boundary_annotation_data_one_to_many_mapping,
     migrate_boundary_annotation_data_one_to_one_mapping,
     migrate_boundary_annotation_data_one_to_one_mapping_float_to_int,
     migrate_boundary_annotation_data_one_to_one_mapping_int_to_float,
 )
-from tests.processing.migration.annotations.test_called_after_migration import (
+from tests.migration.annotations.test_called_after_migration import (
     migrate_called_after_annotation_data_duplicated,
     migrate_called_after_annotation_data_one_to_many_mapping,
     migrate_called_after_annotation_data_one_to_many_mapping__two_classes,
@@ -31,53 +31,53 @@ from tests.processing.migration.annotations.test_called_after_migration import (
     migrate_called_after_annotation_data_one_to_one_mapping__before_splits,
     migrate_called_after_annotation_data_one_to_one_mapping__no_mapping_found,
 )
-from tests.processing.migration.annotations.test_description_migration import (
+from tests.migration.annotations.test_description_migration import (
     migrate_description_annotation_data_duplicated,
     migrate_description_annotation_data_one_to_many_mapping__class,
     migrate_description_annotation_data_one_to_one_mapping__function,
     migrate_description_annotation_data_one_to_one_mapping__parameter,
 )
-from tests.processing.migration.annotations.test_enum_migration import (
+from tests.migration.annotations.test_enum_migration import (
     migrate_enum_annotation_data_duplicated,
     migrate_enum_annotation_data_one_to_many_mapping,
     migrate_enum_annotation_data_one_to_many_mapping__only_one_relevant_mapping,
     migrate_enum_annotation_data_one_to_one_mapping,
 )
-from tests.processing.migration.annotations.test_expert_migration import (
+from tests.migration.annotations.test_expert_migration import (
     migrate_expert_annotation_data__class,
     migrate_expert_annotation_data__function,
     migrate_expert_annotation_data__parameter,
     migrate_expert_annotation_data_duplicated,
 )
-from tests.processing.migration.annotations.test_group_annotation import (
+from tests.migration.annotations.test_group_annotation import (
     migrate_group_annotation_data_duplicated,
     migrate_group_annotation_data_one_to_many_mapping,
     migrate_group_annotation_data_one_to_one_mapping,
     migrate_group_annotation_data_one_to_one_mapping__one_mapping_for_parameters,
 )
-from tests.processing.migration.annotations.test_move_migration import (
+from tests.migration.annotations.test_move_migration import (
     migrate_move_annotation_data_one_to_many_mapping,
     migrate_move_annotation_data_one_to_one_mapping__class,
     migrate_move_annotation_data_one_to_one_mapping__global_function,
     migrate_move_annotation_data_one_to_one_mapping_duplicated,
 )
-from tests.processing.migration.annotations.test_remove_migration import (
+from tests.migration.annotations.test_remove_migration import (
     migrate_remove_annotation_data_duplicated,
     migrate_remove_annotation_data_one_to_many_mapping,
     migrate_remove_annotation_data_one_to_one_mapping,
 )
-from tests.processing.migration.annotations.test_rename_migration import (
+from tests.migration.annotations.test_rename_migration import (
     migrate_rename_annotation_data_duplicated,
     migrate_rename_annotation_data_one_to_many_mapping,
     migrate_rename_annotation_data_one_to_one_mapping,
 )
-from tests.processing.migration.annotations.test_todo_migration import (
+from tests.migration.annotations.test_todo_migration import (
     migrate_todo_annotation_data_duplicated,
     migrate_todo_annotation_data_many_to_many_mapping,
     migrate_todo_annotation_data_one_to_many_mapping,
     migrate_todo_annotation_data_one_to_one_mapping,
 )
-from tests.processing.migration.annotations.test_value_migration import (
+from tests.migration.annotations.test_value_migration import (
     migrate_constant_annotation_data_duplicated,
     migrate_constant_annotation_data_one_to_many_mapping,
     migrate_constant_annotation_data_one_to_one_mapping,
